fix(Player.controller): disable exit time for transitions (#dev)
The "m_HasExitTime" property has been changed from 1 to 0 for several transitions in the Player.controller.
This modification ensures that transitions will no longer wait for the animation to finish before
transitioning to the next state. This change is intended to improve responsiveness and prevent
unnecessary delays in animation playback.