feat(BuiltInStorageConfig): Add LevelToLoad property and reset button (#issue)
Introduces a new property `LevelToLoad` to the `BuiltInStorageConfig` scriptable object.
This property allows for specifying the initial level to load.

Additionally, a new button `ResetLevelToLoad` is added to the Unity Editor.
This button, when clicked, deletes the `_levelToLoad` key from `PlayerPrefs`,
effectively resetting the level to load to its default state.